Fix test project warnings, move kuzu_shared.dll to main output, and move KuzuInternalId into its own file.
